Cleanup, efficiency increases.
VirusMod has been updated to actually work as planned when it comes to how far it spreads. I still have to modify it to degrade also based around the blast resistance of the block it's destroying, but for now it's limited to around 4-10 blocks in each direction of the X Y Z axis per block (Easy maximum of around 10 * 10 * 10 radius, not sure?). There has also been many modifications as to how it works, instead of using a tile entity (which creates alot of over-head), it'll now use the previous blocks metadata to obtain how many left blocks to "eat".