new VirusMod viruses! VirusMod now has two new strains of viruses, the Replacer Virus and the Tool Virus. With this has come a few new changes to how the virus tile entity works (behind the scenes, no real apparent gameplay changes).
The Replacer Virus will use whatever block that is above it as the block to replace what it eats with, currently not fully finished, ran into a slight snag and really wanted to push the commit out.
The Tool Virus will act just like the Eater Virus, however, it will also drop whatever material it minds as a item that you may pickup.

The configuration file has also been edited to be sadly a bit less customizable when it comes to individual viruses, however, in general you may still edit how the virus interacts with the world, how long it lasts, etc. Just some of the options are not virus specific and are grouped accordingly.