Virus changes & efficiency increases The Tool Virus properly works as intended, I was apparently a bit daft and didn't realize what I was really doing when trying to drop the current blocks resource, since it was now set to the VirusStubBlock it wasn't dropping anything. /derp The Replacer Virus acts as intended, whatever is around will be replaced with the block exactly above it in the y direction.
TileEntityVirus has had some performance increases but mostly due to my previous stupidity in how the virus interacts with blocks it's about to replace. It now doesn't iterate through blocks that are the same ID that the virus is replacing the block with. Instant true/false, and if false then checks. Lalala.